ch4/ofi: further refactor init_provider
The code in open_fabric does three functions, find providers, init
multi-nic, and update global limits/settings. Refactor accordingly.
